refactor: extract user menu component
